+++ src/Fable.Cli/CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+
+* [Python] Added `ClassAttributes` attribute to control Python class generation (@dbrattli)
+
 ### Fixed
 
 * [TS/Dart] Fixed optional parameter types (by @ncave)
@@ -105,14 +109,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * [Python] Fix date formatting when repeating a format token more than the known format (example repeating 'd' more than 4 times) (by @MangelMaxime)
 * [JS/TS] Fix #4010: Supports direct nested types when using `jsOptions` (by @MangelMaxime)
 
-    ```fs
+```fs
     let opts =
         jsOptions<Level1> (fun o ->
             o.level2.level3.valueA <- 10
             o.level2.level3.valueB <- 20
             o.topValueA <- 20
         )
-    ```
+```
 
 * [JS/TS] Fix numeric formats (by @MangelMaxime)